fix: paint part in wrong line of Invoice Parts Table
if a Part Number is selected in Invoice Parts Table, then another
TableItem is selected with the mouse first part selected will be painted
into table in newly selected position. the occurred when entering a new
part number and selecting a part already in the table, when selecting a
current part number then selecting another current part already in the
table, or when selecting a current part then selecting the new, blank
entry at the bottom of the table. we added a Paint At Index method to
MyTables and implemented it in all sub-tables.