fix: don't return customerId as invoiceId in DB Query dataMap
I don't think this was ever called because this section of code only
runs if invoiceNum is not -1, ie if we not working with a new invoice.
As we do not have the ability to edit a saved (or cashiered) invoice
then this would never have been a factor. Change code so it is logically
sound and will match other instances of this method implementation in
other DbObjectSavables.

also rearrange some class fields