fix: allow us to save empty text boxes, allowing removal of data from DB
previously, isModified checks would not allow us to save data from a
text box with .getText().length() == 0. this condition would set
.modified to false which would keep us from saving that into the DB. we
now correctly set .modified based on whether or not text box text has
changed compared to the Object as it was when it was first opened. this
allows us to save blank text boxes, effectively removing data from
database.