feat: start JobPart saving, change to InvoicePart saving
previously, we were directly saving InvoicePartTableItems to the
Database. This method of saving InvoiceParts made it clumsy to get
JobParts saved into the Database in a way that made it easy to retrieve
them again (i.e. when opening an existing Repair Order for more work).
So, we have now created an InvoicePart to save into the DB, and we will
have JobPart extend that and @Override key methods.