feat: don't removeAll from InvoicePartTables on PartNumber FOCUS_OUT
This allows us to select an existing Part Number, then immediately
select a new Line's Part Number field without having to click twice.
Implement new & different tests to check whether or not we're editing a
new, blank line so that we can correctly add another new line, ready to
be edited, at the end.

Also, fix a ClassCastException in InvoicePartTable that popped up due to
this change.