fix: NullArgumentException
If an Invoice Part Number field has focus and the application loses
focus (i.e. another application gains focus through a mouse click), that
constitutes a FocusOut event for the Part Number field. This causes the
application to search the Database for the entered part number. If there
is data in the Part Number field that would return more than 1 result
from the Database, the application attempts to open the Part Search
Dialog. In this case there is then no Active Shell associated with the
application (Display.getDefault().getActiveShell() returns null), so
partSearchDialog.open() throws a NullArgumentException.