fix: unable to delete newly created Object
previously, when we sent a new Object to the Table to be painted, it's
DbPk (Database Primary Key) was not set. Therefore, in
DbServices.deleteObject we tried to delete from the Database an entry
with the primary key of -1, which did not exist. Now, we correctly set
the primary key of a newly created Object in the Object itself before it
is painted into the Table.