fix: JobPart.part is null. Change JobPart.setPart to correctly set part
too stringent of checks in JobPart.setPart. When building JobPart in
DbServices we already set partId, partNumber, description etc, so the
conditional in setPart would evaluate to false and we would not set
JobPart.part = part. Therefore JobPart.part would still be null, and
when attempting to open an RO, JobPart.getPart would return null when
attempting to add parts to jobPartTableItem in addPartToPartTableItem
function