fix: ArrayIndexOutOfBoundsException, Add Labor
Job Part and Job Labor buttons are active with no Job in Repair Order.
The Delete buttons do nothing, but Add Labor attempts to add labor, then
calculate the Job total. An ArrayIndexOutOfBoundsException is thrown as
there was no Job in the ROJobTable to modify.

Disable Job Part and Job Labor buttons if there is no Job in the RO.