fix: do not perform DB search unless the user types into search box
Search text boxes start with "Search..." in them. When the user clicks
in the text box, "Search..." is erased allowing the user to type in
their query. If the search box looses focus without anything in it the
string "Search..." is placed back into the search text box. This was
causing the DB to search for "Search...". This did not cause a problem
except with Repair Order search, where at least 1 RO stored in the DB
would be found.

Now we don't search if the search query is "Search..."