fix: disable focus listeners for all Job Details in cashiered RO
If there was nothing entered and saved in the Job Details text boxes,
and therefore the default "Complaints..." or "Resolution..." etc was
there, this caused the default text to disappear when the text box
gained focus, as if to allow the user to edit the text box. Now we
remove the FocusListeners from the text boxes completely disabling this
behavior.