fix: NumberFormatException when no RO part is entered
if an RO Job Part table part number column is clicked into
(ROPartEditorListener activated) then we do not search for a part
number, either through clicking out or pressing [Esc], there will be no
data set in the invoiceDetailText by the InvoicePartEditorListener, and
thusly the new BigDecimal would throw an exception on the empty String.

Skip over that code if the string is empty.