fix: not setting InvoicePartTableItem column data correctly
PartTableItem was only being painted if we were editing a new, blank
TableItem (if TableItem.getData() == null). If we were trying to change
a part line item to a different part, the Part object would be added to
the TableItem but the TableItem would not be painted with the new part's
details. This causes inconsistencies with what is being displayed and
allows other weirdness including being able to set sell quantity higher
than the displayed on hand.