refactor: preparation for checking quantity when part is first added
If a part is added with an on hand quantity of 0, part will be added to
invoice with a sell quantity of 1. The sell quantity is not checked when
initially added to invoice. This is the start of a feature to move the
sell quantity vs on hand quantity check into its own method.