feat: confirm Close and Cancel changes for RO
ask the user to confirm if they Want to Close the RO or Cancel all
changes. Include a check box to let user set whether they want us to ask
again, letting the user decide for themselves the amount of risk they
want to take on vis a vis accidently clicking either of these buttons.