fix: NullPointerException, save new repairOrder to currentRepairOrder
A NullPointerException would occur if a new repair order was saved, and
then saved again without closing it and opening it again. We forgot to
save a newly created repairOrder object into
RepairOrderDialog.currentRepairOrder. Thus, when Save Button is clicked
and we don't find a -1 in roId, we call saveRo(currentRepairOrder), and
there is nothing in currentRepairOrder.

Now we correctly save a repairOrder into currentRepairOrder at the end
of RepairOrderDialog.saveRo.